H4+H6 paisa refactor, vendor escrow breakdown, wallet/purchase flow fixes, Flutter blank screen fix, dead code cleanup
Backend:
- H4: Ledger float64->int64 paisa refactor (wallet/escrow/ledger)
- H6: Full H4 completion - migration 0031+0032 for all 9 remaining tables
- Vendor escrow breakdown endpoint (per-order detail)
- Checkout handler: wallet path uses CalculateSplit for 3-way split
- Wallet deduction before ledger transfer (refund on failure)
- COD dead code cleanup (escrow_cron.go deleted)
- delivery_service: removed premature ledger entries for COD
- Commission: fixed delivery_fee column reference
- Payment handlers: fixed outbox_events columns, SettlementPayload
- Settlement worker: complete transfers array
- Delivery bid persistence, proximity, TTL, stream services
- Order workers, shared money helpers, stock reservations migration
- Admin, ride, analytics, refund handlers updated

Flutter:
- Vendor blank screen fix (removed back buttons from tab screens)
- VendorDashboardScreen: PopScope double-back-to-exit
- Product details: 434 lines restored, B1/B2/B3/B7/B9 fixes
- Checkout: OrderSuccessScreen for all payment methods
- Customer wallet, dashboard, order screens updated
- Vendor add product: dead code removed, ApiClient fix
- New: saved cards, my orders, security settings screens

+++ backend/go-services/internal/escrow/models.go
@@ -22,7 +22,7 @@ type EscrowHold struct {
 	ID               uuid.UUID    `json:"id"`
 	OrderTrackingID  string       `json:"order_tracking_id"`
 	VendorTrackingID string       `json:"vendor_tracking_id"`
-	Amount           float64      `json:"amount"`
+	Amount           int64        `json:"amount"` // paisa
 	Status           EscrowStatus `json:"status"`
 	HoldUntil        time.Time    `json:"hold_until"`
 	ReleasedAt       *time.Time   `json:"released_at,omitempty"`
